Split the CSE machinery into its own module.
