+++ dhall-to-cabal.cabal
@@ -43,6 +43,7 @@ library
         CabalToDhall
         DhallLocation
         DhallToCabal
+        DhallToCabal.FactorType
         DhallToCabal.Util
     hs-source-dirs: lib
     other-modules:
@@ -68,6 +69,7 @@ library
         contravariant ^>=1.4 || ^>=1.5,
         dhall ^>=1.24.0,
         filepath ^>=1.4,
+        microlens ^>=0.1.0.0 || ^>=0.2.0.0 || ^>=0.3.0.0 || ^>=0.4.0.0,
         text ^>=1.2,
         transformers ^>=0.5.2,
         vector ^>=0.12
